refactor: money 객체 생성
LottoMachine 에 입력되는 돈에 대한 객체 신규 추가
